# scrape: Add realistic data case for scrape loop append bench.