Helps with https://github.com/prometheus/prometheus/issues/14668

In the past we only benchmark one metric family with only gauges, which is not realistic. I kept old scenario so we can optimize that case, but added realistic dataset with tons of histograms and 59 metric families. Interestingly in such scenario textparse is not so slow (allocation heavily though).

```
benchstat -col /fmt current=append-v1.txt
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/scrape
cpu: Apple M1 Pro
                                       │   PromText   │               OMText               │             PromProto              │
                                       │    sec/op    │    sec/op     vs base              │   sec/op     vs base               │
ScrapeLoopAppend/data=1Fam1000Gauges-2   280.4µ ± 10%   265.1µ ± 10%  -5.45% (p=0.041 n=6)   447.0µ ± 2%  +59.41% (p=0.002 n=6)
ScrapeLoopAppend/data=59FamsAllTypes-2   168.9µ ±  3%   166.4µ ±  1%  -1.51% (p=0.041 n=6)   159.0µ ± 3%   -5.90% (p=0.002 n=6)
geomean                                  217.7µ         210.0µ        -3.50%                 266.6µ       +22.48%

                                       │   PromText   │               OMText                │               PromProto               │
                                       │     B/op     │     B/op       vs base              │     B/op       vs base                │
ScrapeLoopAppend/data=1Fam1000Gauges-2   178.7Ki ± 8%   170.5Ki ± 10%       ~ (p=0.394 n=6)    665.7Ki ± 2%  +272.54% (p=0.002 n=6)
ScrapeLoopAppend/data=59FamsAllTypes-2   74.02Ki ± 1%   72.87Ki ±  1%       ~ (p=0.065 n=6)   169.16Ki ± 1%  +128.54% (p=0.002 n=6)
geomean                                  115.0Ki        111.5Ki        -3.09%                  335.6Ki       +191.79%

                                       │  PromText  │              OMText              │               PromProto                │
                                       │ allocs/op  │ allocs/op   vs base              │  allocs/op    vs base                  │
ScrapeLoopAppend/data=1Fam1000Gauges-2   11.00 ± 0%   12.00 ± 0%  +9.09% (p=0.002 n=6)   8012.00 ± 0%  +72736.36% (p=0.002 n=6)
ScrapeLoopAppend/data=59FamsAllTypes-2   18.00 ± 0%   19.00 ± 0%  +5.56% (p=0.002 n=6)   1693.00 ± 0%   +9305.56% (p=0.002 n=6)
geomean                                  14.07        15.10       +7.31%                  3.683k       +26073.77%

```


Profiles:
* [MEM 1000 gauges proto](https://pprof.me/eef12446746ec8b501e6197e8a679d3e/?profileType=profile%253Aalloc_objects%253Acount%253Aspace%253Abytes&color_by=filename)
* [CPU 1000 gauges proto](https://pprof.me/b4cf5a247a0678af227bcf8da7038d28/?profileType=profile%253Acpu%253Ananoseconds%253Acpu%253Ananoseconds%253Adelta&color_by=filename)
* [MEM 59 MFs](https://pprof.me/3fe621bcab564605cee2bb74d1d2db3b/?profileType=profile%253Aalloc_objects%253Acount%253Aspace%253Abytes&color_by=filename)